rag-engine/requirements.txt
flask
flask-cors
faiss-cpu
requests
python-dotenv
numpy